黃振瑋.新增: Admin的interface
